chore: Change network on dapp method request
